requirements_merged.txt
# ==================================================
# MERGED TRADING BOT REQUIREMENTS
# Original + Institutional Grade System
# Compatible with Python 3.13
# ==================================================

# Core Python Requirements
python-telegram-bot==20.7
tensorflow==2.20.0rc0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
ta-lib==0.4.28
requests==2.31.0
websocket-client==1.7.0
python-socketio==5.10.0
aiohttp==3.9.1
schedule==1.2.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
yfinance==0.2.28
ccxt==4.1.77
python-dotenv==1.0.0
psutil==5.9.6
pytz==2024.1
joblib==1.3.2
sqlalchemy==2.0.25
json5==0.9.14
beautifulsoup4==4.12.2
cryptography==42.0.2
xgboost==2.0.3
optuna==3.5.0
scipy==1.12.0
textblob==0.17.1
feedparser==6.0.10

# Professional Data Providers
polygon-python-client==1.13.2
alpha-vantage==2.3.1
iex-python==0.3.0
quandl==3.7.0

# Financial Libraries
quantlib==1.32
pyfolio==0.9.2
empyrical==0.5.5
bt==0.2.9

# High-Performance Computing
numba==0.59.0
cython==3.0.8
numexpr==2.9.0
bottleneck==1.3.7

# Database and Caching
redis==5.0.1
influxdb-client==1.38.0
clickhouse-driver==0.2.6
pymongo==4.6.0

# Message Queues and Streaming
apache-kafka-python==2.0.2
pika==1.3.2
celery==5.3.4

# Monitoring and Logging
prometheus-client==0.19.0
sentry-sdk==1.39.2
loguru==0.7.2
structlog==23.2.0

# Network and Security
paramiko==3.4.0
keyring==24.3.0
pycryptodome==3.19.0

# Time Series Analysis
statsmodels==0.14.1
arch==6.2.0
tslearn==0.6.2
prophet==1.1.5
sktime==0.25.0

# Alternative Data
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.7
textstat==0.7.3
vaderSentiment==3.3.2